+++ src/main/scala/com/cicroomapi/models/RoomsTable.scala
@@ -1,0 +1,12 @@
+package com.cicroomapi.models
+
+import org.scalatra._
+import slick.driver.PostgresDriver.api._
+
+class RoomsTable(tag: Tag) extends Table[(Int, String)](tag, "rooms") {
+  
+  def id = column[Int]("id", O.PrimaryKey, O.AutoInc)
+  def name = column[String]("name")
+
+  def * = (id, name)
+}
